Updated the game progression mechanics. Added Enums fir location and weather to ensure consistency.

+++ internal/config/information/infotools/inttobar.go
@@ -2,13 +2,13 @@ package infotools
 
 import "strings"
 
-func IntToBar(total int) string {
+func IntToBar(total float64) string {
 	filledBars := total / 10
 	barval := "|"
 	if filledBars == 10 {
 		barval = "█"
 	}
-	bars := strings.Repeat(barval, filledBars)
+	bars := strings.Repeat(barval, int(filledBars))
 
 	return bars
 }
